+++ BookSearch/src/Book.java
@@ -1,0 +1,13 @@
+
+
+public class Book {
+    private int bookId;
+
+    public Book(int bookId) {
+        this.bookId = bookId;
+    }
+
+    public int getBookId() {
+        return bookId;
+    }
+}
