--- BookSearch/src/com/example/booksearch/Book.java
@@ -1,13 +1,0 @@
-package com.example.booksearch;
-
-public class Book {
-    private int bookId;
-
-    public Book(int bookId) {
-        this.bookId = bookId;
-    }
-
-    public int getBookId() {
-        return bookId;
-    }
-}
